+++ src/lazyLoad/next_js/next_GetLazyResourcesWebpackJs.ts
@@ -29,6 +29,7 @@ const next_GetLazyResourcesWebpackJs = async (
 ): Promise<string[] | any> => {
     const browser = await puppeteer.launch({
         headless: true,
+        args: process.env.IS_DOCKER === "true" ? ["--no-sandbox"] : [],
     });
 
     const page = await browser.newPage();